requirement.txt
absl-py==1.4.0
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
AMFM-decompy==1.0.11
antlr4-python3-runtime==4.8
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.1
bitarray==2.7.3
bleach==6.0.0
byted-torch @ file:///opt/tiger/byted_torch_workspace/byted_torch
byted-wandb==0.13.27
bytedance-context==0.7.1
bytedance-metrics==0.5.0
bytedbackgrounds==0.0.6
byteddps==0.1.2
bytedenv==0.6.0
bytedlego-ops @ file:///bytedlego_ops-1.0.0.383-py2.py3-none-any.whl
bytedlogger==0.14.6
bytedmemfd==0.1
bytedmetrics==0.10.2
bytedservicediscovery==0.16.0
bytedtcc==1.4.2
bytedtos==1.1.11
bytedtrace==0.1.7
bytedztijwthelper==0.0.20
bytedztispiffe==0.0.9
byteps @ http://luban-source.byted.org/repository/scm/aml.arnold.byteps_1.0.0.651.tar.gz
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
cryptography==38.0.4
cycler==0.11.0
Cython==0.29.34
datasets==2.11.0
debugpy==1.6.7
decorator==5.1.1
deepspeed==0.8.3
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
Distance==0.1.3
distro-info==0.21
dnspython==2.3.0
docker-pycreds==0.4.0
dtw-python==1.3.0
editdistance==0.6.2
einops==0.6.0
entrypoints==0.4
evaluate==0.4.0
exceptiongroup==1.1.1
fastjsonschema==2.16.3
ffmpeg==1.4
filelock==3.10.7
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
g2p-en==2.1.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
hjson==3.1.0
horovod==0.25.0
huggingface-hub==0.13.3
hydra-core==1.0.7
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflect==6.0.2
iniconfig==2.0.0
iotop==0.6
ipaddress==1.0.23
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jiwer==3.0.1
joblib==1.2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.6
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.36.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
mock==5.0.1
msgpack==0.6.2
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
ninja==1.11.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
npy-append-array==0.9.13
numba==0.53.0
numpy==1.21.5
oauthlib==3.2.2
omegaconf==2.0.6
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycurl==7.43.0.2
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.30.4
PyJWT==1.7.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypinyin==0.48.0
pyrsistent==0.19.3
pytest==7.2.2
python-apt==1.8.4.3
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2023.3
pyworld==0.3.2
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
sacrebleu==2.3.1
sacremoses==0.0.53
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==1.15.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.2
submitit==1.4.5
subword-nmt==0.3.8
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
tensorrt @ file:///tensorrt-8.0.3.4-cp37-none-linux_x86_64.whl
termcolor==2.2.0
terminado==0.17.1
thai-segmenter==0.4.1
threadpoolctl==3.0.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch @ file:///tmp/pip-install-804t2osu/byted-torch_78874ef99d1540fd821fcec4fcb49d56/src/byted_torch/torch-1.10.0-cp37-cp37m-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.4.0
unattended-upgrades==0.1
Unidecode==1.3.6
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wrapt==1.14.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.13.0
